Switch from yarn to npm for now
This is because right now yarn does not properly normalize the file permission across installed packages. This means that, if a package contains a file with permission that won't allow it to be read later on (e.g. install package global so it's owned by root but it doesn't has read access for others), things will break. There is a PR to add permission normalization but it has not been merged to Yarn yet.